lightningd: clamp estimatefees results to a sanity ceiling
Nothing bounded the feerates we accept from the `estimatefees` plugin:
`parse_feerate_ranges` only checks that block counts ascend and that
rates do not increase with block count, and the conversion loop only
applies a floor.  A broken or misbehaving fee source could therefore
push an arbitrary u32 into every downstream feerate calculation.

Clamp both the reported floor and each individual rate to
FEERATE_CEILING once they have been converted to perkw, and log loudly
when we do.  We clamp rather than reject because `bitcoin_plugin_error`
is fatal(), and because discarding the estimate entirely is worse than
capping it: with no rates at all `feerate_max` falls into its "unknown"
path and returns UINT_MAX.

The ceiling is deliberately not a policy limit.  It sits at 4000 sat/vB,
several times the highest the chain has ever seen, so that it only ever
fires on a broken fee source.  Picking a tighter bound would mean
underestimating during genuine congestion, which is exactly when our
unilateral close and HTLC timeout transactions need to confirm.

The conversion has to be widened for the ceiling to see anything at all.
feerate_from_style() computed (feerate + 3) / 4 on a u32, so the top
three perkb values wrapped to 0 or 1 perkw, slipped under the ceiling,
and were then raised to the floor, hiding the whole thing.  That is also
the more dangerous direction, since it underpays our own unilateral
close, and it applies equally to operator input like
`feerate="4294967295perkb"`.  feerate_to_style() right below already
widened for the same reason; the asymmetry was an oversight.

Changelog-Fixed: lightningd: feerate estimates from the Bitcoin backend are now clamped to a sanity ceiling, so a misbehaving `estimatefees` plugin cannot feed absurd feerates into the rest of the daemon.
Changelog-Fixed: lightningd: a feerate given in perkb near the top of the 32-bit range no longer wraps to a near-zero perkw feerate.

+++ bitcoin/feerate.c
@@ -12,8 +12,15 @@ u32 feerate_from_style(u32 feerate, enum feerate_style style)
 		return feerate;
 	case FEERATE_PER_KBYTE:
 		/* Everyone uses satoshi per kbyte, but we use satoshi per ksipa
-		 * (don't round down to zero though)! */
-		return (feerate + 3) / 4;
+		 * (don't round down to zero though)!
+		 *
+		 * Widen before rounding up: on a u32 the +3 wraps for the top
+		 * three values, turning an absurd feerate into 0 or 1 perkw.
+		 * That is the dangerous direction (it underpays our unilateral
+		 * close), and it slips under every bound we check afterwards,
+		 * since those are applied to the converted value.  The result
+		 * always fits a u32: (UINT_MAX + 3) / 4 < UINT_MAX. */
+		return ((u64)feerate + 3) / 4;
 	}
 	abort();
 }
